requirements.txt
certifi==2020.12.5
chardet==4.0.0
idna==2.10
Mastodon.py==1.5.1
oauthlib==3.1.0
python-dateutil==2.8.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
twython==3.8.2
urllib3==1.26.4
